requirements.txt
Django~=3.0.3
mysqlclient~=1.4.6
django-crispy-forms~=1.8.1
pillow~=7.0.0
sorl-thumbnail~=12.6.3
django-filter~=2.2.0
django-widget-tweaks~=1.4.8
django-phonenumber-field[phonenumbers]~=4.0.0